feat: exercise 4.22

+++ part4/blog-list/controllers/blogsController.js
@@ -1,15 +1,17 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const middleware = require('../utils/middleware')
 require('express-async-errors') // no need to use try and catch with this library
 
+
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({}).populate('user')
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1})
   response.json(blogs)
 })
 
 blogsRouter.get('/:id', async (request, response) => { 
-  const blog = await Blog.findById(request.params.id)
+  const blog = await Blog.findById(request.params.id).populate('user', { username: 1, name: 1, id: 1})
   if(blog){
     response.json(blog)
   }else {
@@ -17,8 +19,8 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const user = await User.findOne()
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const user = request.user
 
   const blog = new Blog({
     title: request.body.title,
@@ -27,14 +29,27 @@ blogsRouter.post('/', async (request, response) => {
     likes: request.body.likes || 0,
     user: user.id
   })
+  
   const result = await blog.save()
   user.blogs = user.blogs.concat(blog._id)
   await user.save()
   response.status(201).json(result)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  const result = await Blog.findByIdAndDelete(request.params.id)
+blogsRouter.delete('/:id', middleware.userExtractor ,async (request, response) => { 
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if(!blog){
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if(blog.user.toString() !== user.id){
+    return response.status(401).json({ error: 'unauthorized'})
+  }
+
+  await Blog.findByIdAndDelete(request.params.id)
+  console.log("deleted succesfully")
   response.status(204).end()
 })
